requirements.txt
fastapi
starlette
uvicorn[standard]
python-dotenv
requests
google-generativeai
openai
gne
playwright==1.48.0
newspaper4k
lxml_html_clean
nltk
jieba
notion-client
SQLAlchemy
PyMySQL
slowapi[redis]
redis
cachetools
youtube-transcript-api